Removed all the static casts

+++ include/nabla4_structured_torus.hpp
@@ -105,8 +105,7 @@ class nabla4_structured_torus {
         // " longitude: " << longitude << " longitude_dim: " << longitude_dim << " latitude_dim: " << latitude_dim << "
         // latitude_m1: " << latitude_m1 << std::endl;
         e2c2v_ret[2] =
-            ((((latitude == 0) * static_cast<size_t>((2 * longitude_dim - latitude_dim) / 2)) + longitude + 1) %
-                longitude_dim) *
+            ((((latitude == 0) * ((2 * longitude_dim - latitude_dim) / 2)) + longitude + 1) % longitude_dim) *
                 latitude_dim +
             latitude_m1;
         e2c2v_ret[3] =
@@ -146,11 +145,10 @@ class nabla4_structured_torus {
         std::size_t longitude_m1) {
         std::array<ARRAY_TYPE, 4> e2c2v_ret{};
         e2c2v_ret[0] = parent_vertex;
-        e2c2v_ret[1] = ((((latitude == latitude_dim - 1) * static_cast<size_t>(latitude_dim / 2)) + longitude_dim +
-                            longitude - 1) %
-                           longitude_dim) *
-                           latitude_dim +
-                       latitude_p1;
+        e2c2v_ret[1] =
+            ((((latitude == latitude_dim - 1) * (latitude_dim / 2)) + longitude_dim + longitude - 1) % longitude_dim) *
+                latitude_dim +
+            latitude_p1;
         e2c2v_ret[2] =
             ((((latitude == latitude_dim - 1) * (latitude_dim / 2)) + longitude) % longitude_dim) * latitude_dim +
             latitude_p1;
@@ -182,14 +180,14 @@ class nabla4_structured_torus {
         const auto E2ECV_1 = edge_index * 4 + 1;
         const auto E2ECV_2 = edge_index * 4 + 2;
         const auto E2ECV_3 = edge_index * 4 + 3;
-        double nabv_tang_wp = static_cast<double>(u_vert[E2C2V_0][k_index]) * primal_normal_vert_v1[E2ECV_0] +
-                              static_cast<double>(v_vert[E2C2V_0][k_index]) * primal_normal_vert_v2[E2ECV_0] +
-                              static_cast<double>(u_vert[E2C2V_1][k_index]) * primal_normal_vert_v1[E2ECV_1] +
-                              static_cast<double>(v_vert[E2C2V_1][k_index]) * primal_normal_vert_v2[E2ECV_1];
-        double nabv_norm_wp = static_cast<double>(u_vert[E2C2V_2][k_index]) * primal_normal_vert_v1[E2ECV_2] +
-                              static_cast<double>(v_vert[E2C2V_2][k_index]) * primal_normal_vert_v2[E2ECV_2] +
-                              static_cast<double>(u_vert[E2C2V_3][k_index]) * primal_normal_vert_v1[E2ECV_3] +
-                              static_cast<double>(v_vert[E2C2V_3][k_index]) * primal_normal_vert_v2[E2ECV_3];
+        double nabv_tang_wp = u_vert[E2C2V_0][k_index] * primal_normal_vert_v1[E2ECV_0] +
+                              v_vert[E2C2V_0][k_index] * primal_normal_vert_v2[E2ECV_0] +
+                              u_vert[E2C2V_1][k_index] * primal_normal_vert_v1[E2ECV_1] +
+                              v_vert[E2C2V_1][k_index] * primal_normal_vert_v2[E2ECV_1];
+        double nabv_norm_wp = u_vert[E2C2V_2][k_index] * primal_normal_vert_v1[E2ECV_2] +
+                              v_vert[E2C2V_2][k_index] * primal_normal_vert_v2[E2ECV_2] +
+                              u_vert[E2C2V_3][k_index] * primal_normal_vert_v1[E2ECV_3] +
+                              v_vert[E2C2V_3][k_index] * primal_normal_vert_v2[E2ECV_3];
         z_nabla4_e2_wp[edge_index][k_index] =
             4.0 * ((nabv_norm_wp - 2.0 * z_nabla2_e[edge_index][k_index]) *
                           (inv_vert_vert_length[edge_index] * inv_vert_vert_length[edge_index]) +
